Выполнил задачу
